// src/app/models/timesheet/tsLineListContact.interface.ts
export class TimesheetLine {
    // ActivityName:any; 
    // CategoryId:string;
    // CategoryName:any;
    // IsDeleted:Number;
    // LineNum:Number;
    // ProjActivityNumber:string;
    // ProjId:string;
    // ProjName:string;
    // TimesheetLineDateList:TimesheetLineDateList[];
    // TotalHrs:any;

    TimesheetCode: any;
    LineNumber: any;

    ProjId: any;
    ProjDescription: any;

    ProjActivityId: any;
    ProjActivityDescription: any;

    CategoryId: any;
    CategoryDescription: any;

    Hours1: any;
    Hours2: any;
    Hours3: any;
    Hours4: any;
    Hours5: any;
    Hours6: any;
    Hours7: any;


    EnterHours1: any;
    EnterHours2: any;
    EnterHours3: any;
    EnterHours4: any;
    EnterHours5: any;
    EnterHours6: any;
    EnterHours7: any;


    InternalComment1: any;
    InternalComment2: any;
    InternalComment3: any;
    InternalComment4: any;
    InternalComment5: any;
    InternalComment6: any;
    InternalComment7: any;
    
    ExternalComment1: any;
    ExternalComment2: any;
    ExternalComment3: any;
    ExternalComment4: any;
    ExternalComment5: any;
    ExternalComment6: any;
    ExternalComment7: any;
    Number: any;

    IsDeleted:boolean;
    toggle:boolean;

    DataArea:any;
    ERPRecID:any;
    constructor(){

    }
}